Conflicts:
	src/core/tools/common/simple_logger.cpp

+++ tutorial/counter.replication/counter.server.impl.cpp
@@ -168,8 +168,7 @@ namespace dsn {
 
             // TODO: should use async write instead
             char name[256];
-            sprintf(name, "%s/checkpoint.%lld", data_dir().c_str(),
-                static_cast<long long int>(last_committed_decree()));
+            sprintf(name, "%s/checkpoint.%" PRId64, data_dir().c_str(), last_committed_decree());
             std::ofstream os(name);
 
             uint32_t count = (uint32_t)_counters.size();
